Доработки от 01.12.2024
Исправил всё по пожеланиям от 22.11.204
