+++ Homeworks/Base/src/PromoCodeFactory.Core/Domain/Administration/Employee.cs
@@ -5,12 +5,12 @@ namespace PromoCodeFactory.Core.Domain.Administration
 {
     public class Employee : BaseEntity
     {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
+        public required string FirstName { get; set; }
+        public required string LastName { get; set; }
 
         public string FullName => $"{FirstName} {LastName}";
 
-        public string Email { get; set; }
+        public required string Email { get; set; }
 
         public required List<Role> Roles { get; set; }
