+++ Homeworks/Base/src/PromoCodeFactory.Core/Domain/Administration/Role.cs
@@ -2,8 +2,8 @@
 {
     public class Role : BaseEntity
     {
-        public string Name { get; set; }
+        public required string Name { get; set; }
 
-        public string Description { get; set; }
+        public required string Description { get; set; }
     }
 }
